Refactor tests and enhance ListenableBindHandlerAdapter
Refactored test assertions for server port into a shared AbstractTest utility. Improved ListenableBindHandlerAdapter to support onCreate via MethodHandle for compatibility with different Spring Boot versions. Updated and cleaned up related test classes for consistency and maintainability.

+++ microsphere-spring-boot-core/src/test/java/io/microsphere/spring/boot/context/properties/source/util/ConfigurationPropertyUtilsTest.java
@@ -49,7 +49,7 @@ class ConfigurationPropertyUtilsTest {
                 assertEquals(prefix, getPrefix(name, context));
             }
         });
-        assertEquals(12345, serverProperties.getPort());
+        assertEquals((Integer) 12345, serverProperties.getPort());
     }
 
     @Test
